+++ app/controllers/projects_controller.rb
@@ -38,19 +38,17 @@ class ProjectsController < ApplicationController
     end
   end
 
+  # GET /projects/1/subject_report
+  # GET /projects/1/subject_report.js
   def subject_report
     current_user.pagination_set!('subjects', params[:subjects_per_page].to_i) if params[:subjects_per_page].to_i > 0
     @statuses = params[:statuses] || ['valid', 'pending', 'test']
     @subjects = @project.subjects.where(site_id: current_user.all_viewable_sites.pluck(:id), status: @statuses).order('subject_code').page(params[:page]).per( current_user.pagination_count('subjects') )
     @designs = @project.designs.order('name')
-
-    respond_to do |format|
-      format.html # subject_report.html.erb
-      format.json { render json: @project }
-      format.js { render 'subject_report' }
-    end
   end
 
+  # GET /projects/1/splash
+  # GET /projects/1/splash.js
   def splash
     @projects = current_user.all_viewable_and_site_projects.order(:name).page(params[:page]).per( 8 ) # current_user.pagination_count('projects') )
     redirect_to @projects.first if @projects.total_count == 1
@@ -72,18 +70,12 @@ class ProjectsController < ApplicationController
     file_pdf_location = @design.latex_report_new_file_location(current_user, orientation, @report_title, @report_subtitle, @report_caption, @percent, @table_header, @table_body, @table_footer)
 
     if File.exists?(file_pdf_location)
-      File.open(file_pdf_location, 'r') do |file|
-        file_name = @report_title.gsub(' vs. ', ' versus ').gsub(/[^\da-zA-Z ]/, '')
-        send_file file, filename: "#{file_name} #{Time.now.strftime("%Y.%m.%d %Ih%M %p")}.pdf", type: "application/pdf", disposition: "inline"
-      end
+      file_name = @report_title.gsub(' vs. ', ' versus ').gsub(/[^\da-zA-Z ]/, '')
+      send_file file_pdf_location, filename: "#{file_name} #{Time.now.strftime("%Y.%m.%d %Ih%M %p")}.pdf", type: "application/pdf", disposition: "inline"
     else
       render text: "PDF did not render in time. Please refresh the page."
     end
   end
-
-  # def report
-  #   setup_report
-  # end
 
   def remove_file
     @project.remove_logo!
@@ -184,68 +176,6 @@ class ProjectsController < ApplicationController
         # Will automatically generate a site if the project has no site
         :site_name
       )
-    end
-
-    def setup_report
-      @sheet_before = parse_date(params[:sheet_before])
-      @sheet_after = parse_date(params[:sheet_after])
-
-      @by = ["week", "month", "year"].include?(params[:by]) ? params[:by] : "month" # "month" or "year"
-      @percent = ['none', 'row', 'column'].include?(params[:percent]) ? params[:percent] : 'none'
-      @statuses = params[:statuses] || ['valid']
-
-      sheet_scope = current_user.all_viewable_sheets.with_project(@project.id)
-      sheet_scope = sheet_scope.sheet_after(@sheet_after) unless @sheet_after.blank?
-      sheet_scope = sheet_scope.sheet_before(@sheet_before) unless @sheet_before.blank?
-
-      sheet_scope = sheet_scope.with_subject_status(@statuses)
-
-      min = sheet_scope.collect{|s| s.created_at.to_date}.min || Date.today
-      max = sheet_scope.collect{|s| s.created_at.to_date}.max || Date.today
-
-      # @ranges = [{ name: "2012", start_date: "2012-01-01", end_date: "2012-12-31" }, { name: "2013", start_date: "2013-01-01", end_date: "2013-12-31" }]
-      @ranges = []
-
-      case @by when "week"
-        current_cweek = min.cweek
-        (min.year..max.year).each do |year|
-          (current_cweek..Date.parse("#{year}-12-28").cweek).each do |cweek|
-            start_date = Date.commercial(year,cweek) - 1.day
-            end_date = Date.commercial(year,cweek) + 5.days
-            @ranges << { name: "Week #{cweek}", tooltip: "#{year} #{start_date.strftime("%m/%d")}-#{end_date.strftime("%m/%d")} Week #{cweek}", start_date: start_date, end_date: end_date }
-            break if year == max.year and cweek == max.cweek
-          end
-          current_cweek = 1
-        end
-      when "month"
-        current_month = min.month
-        (min.year..max.year).each do |year|
-          (current_month..12).each do |month|
-            start_date = Date.parse("#{year}-#{month}-01")
-            end_date = Date.parse("#{year}-#{month}-01").end_of_month
-            @ranges << { name: "#{Date::ABBR_MONTHNAMES[month]} #{year}", tooltip: "#{Date::MONTHNAMES[month]} #{year}", start_date: start_date, end_date: end_date }
-            break if year == max.year and month == max.month
-          end
-          current_month = 1
-        end
-      when "year"
-        @ranges = (min.year..max.year).collect{|year| { name: year.to_s, tooltip: year.to_s, start_date: Date.parse("#{year}-01-01"), end_date: Date.parse("#{year}-12-31") }}
-      end
-
-      between = if @sheet_after.blank? and @sheet_before.blank?
-        "All Sheets"
-      elsif @sheet_after.blank?
-        "Date before #{@sheet_before.strftime("%b %d, %Y")}"
-      elsif @sheet_before.blank?
-        "Date after #{@sheet_after.strftime("%b %d, %Y")}"
-      else
-        "Date between #{@sheet_after.strftime("%b %d, %Y")} and #{@sheet_before.strftime("%b %d, %Y")}"
-      end
-
-      @report_title = 'Design vs. Sheet Creation Date'
-      @report_caption = "#{@project.name}"
-
-      @sheets = sheet_scope
     end
 
 end